# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple password generator API built with Go and Gin.",
        "title": "EBVN Bookmark API",
        "contact": {},
        "version": "1.4"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/self/info": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the authenticated user's profile information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_internal_dto.SuccessResponse-github_com_HadesHo3820_ebvn-golang-course_internal_model_User"
                        }
                    },
                    "401": {
                        "description": "Invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_pkg_response.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_pkg_response.Message"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the authenticated user's profile information (display name and/or email)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update user profile",
                "parameters": [
                    {
                        "description": "Update profile request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_user.updateSelfInfoReqBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_internal_dto.SuccessResponse-any"
                        }
                    },
                    "400": {
                        "description": "No data provided for update",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_pkg_response.Message"
                        }
                    },
                    "401": {
                        "description": "Invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_pkg_response.Message"
                        }
                    },
                    "404": {
                        "description": "User does not exist",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_pkg_response.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_pkg_response.Message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_HadesHo3820_ebvn-golang-course_internal_dto.Metadata": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer",
                    "example": 1
                },
                "first_page": {
                    "type": "integer",
                    "example": 1
                },
                "last_page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 10
                },
                "total_records": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "github_com_HadesHo3820_ebvn-golang-course_internal_dto.SuccessResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_internal_dto.Metadata"
                }
            }
        },
...
        "github_com_HadesHo3820_ebvn-golang-course_internal_dto.SuccessResponse-github_com_HadesHo3820_ebvn-golang-course_internal_model_User": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_internal_model.User"
                },
                "message": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/github_com_HadesHo3820_ebvn-golang-course_internal_dto.Metadata"
                }
            }
        },
...
        "github_com_HadesHo3820_ebvn-golang-course_internal_model.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_HadesHo3820_ebvn-golang-course_pkg_response.Message": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "Details contains a list of specific error messages or additional information, if any.\nIf empty, this field is omitted from the JSON response."
                },
                "message": {
                    "description": "Message is a brief summary of the response (e.g., \"Input error\").",
                    "type": "string"
                }
            }
        },
...
        "internal_handler_user.updateSelfInfoReqBody": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "description": "`validate:\"omitempty,email\"` allow empty email (only validates format when a value is provided)",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your Bearer token in the format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}